Add pause support to Music class
If the scene tree is paused, engine time keeps ticking.
I resolve that by adjusting our anchor point based on the amount of time spent paused.
I also adjust it so the current offset stays mostly correct, by using the difference between the stored offset and a newly measured one.
I'm using the middle of the current audio mix because it aligns well with how we calculate offsets.
This method is usually around +-3ms off (can't get better without looking into the future), but because it goes in both directions, that averages out to 0.