Create song.gd
This class represents a song and all of its charts.
Right now it doesn't store any info about the charts; I'll do that when I implement a song select screen.
Title and artist are obvious, offset means the amount of time from the start of the audio file to beat zero (useful for songs with some lead in time).