Add Timing and Layer classes
And add them to Game in this commit because they're coupled (perhaps I'll change that in the future, though it wouldn't make much sense).
Timing handles conversion between beat and time based on the tempo changes in the song and the starting offset.
Layer handles conversion between time and position/scale. There can be multiple layers to allow independent speeds for different notes. The position of a note relative to the layer it's on determines where it ends up on the screen.
Timing and speed use an integral approximation, zoom doesn't need to because it doesn't accumulate anything.
For speed curves I generate 20 constant speeds per second of curve; it's smooth enough and performs well.