Setup and update timing and layers in Game
For this game I'm going with 8 layers, should be plenty.
Update should receive time from the Music class.
Timing.time_audio is adjusted for audio latency and will be used for input logic.
Timing.time_video is adjusted for video latency and will be used for rendering.
Layers are used for rendering, so they use video time.