Remove song/chart/timing/layers from Game
Just because taronuke's game puts it in an autoload doesn't mean I have to.
I think it makes more sense to store them in the scenes that use them, and pass them along to objects in need.
I'm planning on using Song and Timing in the options and drift scenes.
For example if I wanted to put the options on top of the gameplay while it's paused, I don't want to have to overwrite the song/timing used by the gameplay.
The neat thing about version control is that I can always undo this change if I end up regretting it.