Change Song and Chart to Resource types
To make this work in the editor, their subclasses (Tempo/Speed/Zoom/Note) have to be in the global scope, so they get their own files.
This is kind of a lot of files, so I made new folder for them.
Song stores a list of charts now, as well as an AudioStream; this way it has all of the relevant data we'll need for song/chart select and gameplay.