Undo the last two changes to SyncManager
I changed my mind, because if I did this with Layer it would not work due to export var setters being called before ready.
I need the instructions to flow down in a specific way, and doing that manually through a script is easier than doing it the Godot way.
I also came back around on the function names, for reasons I'm too lazy to write down.