Use node relation in SyncManager and ViewLayer
On the one hand, coupling is bad.
On the other hand, using a view layer without a sync manager would be silly, as would using a sync manager without a music player.
I'm sticking with setup and update methods rather than trying to use ready or process, because the order of operations is important, and it's easier to do that manually.