Fix MusicPlayer pause before play
If you paused the tree before the sound started playing, it would not handle things properly.
I solve this by storing the time at which we paused and comparing to the time at which we unpause.
This is only done when the sound is not playing, otherwise I still use the mix center approach.
I'm also using it for get_time_engine while paused, instead of checking whether the audio is playing in general, because engine time needs to keep changing before (in the negative) and after (to keep the visuals moving) the audio plays.
I can't just check whether the tree is paused, because the process mode could flip things.
Don't use stream_paused directly, only pause by pausing the scene tree, so it can receive the pause/unpause notifications.
For get_time_player I use not playing instead of stream_paused, because I don't want mix time affecting the output while the audio isn't playing.
Neither of these really matter, because I'm not calling these functions while the tree is paused anyway; it's there just in case.