Change window start resolution
A target resolution of 1920x1080 is great, but having the window start at that size on a screen of the same resolution is awkward, as you can't access the title bar.
I'll start the game with a much smaller window of 1152x648 (godot's default viewport size), because it fits comfortably even on a 1280x720 screen (the smallest you'd realistically support nowadays).